# Conflicts:
#	examples/cpal_test.rs
#	src/config.rs
#	src/info.rs
#	src/lib.rs
#	src/mic.rs
#	src/mic/converters.rs
#	src/mic/mic_config.rs
#	src/mic/mic_cpal.rs
#	src/mic/process_audio.rs
#	src/mic/resampler.rs
#	src/model.rs
#	src/model_tests.rs
#	src/oww.rs
#	src/oww/audio.rs
#	src/oww/oww_model.rs
#	src/save.rs
#	src/tests.rs

+++ src/mic.rs
@@ -10,6 +10,6 @@ pub const RMS_BUFFER_SIZE: usize = 16; // 1 secs+ buffer size
 
 pub const EXPECTED_CHUNK_SIZE: usize = 512;
 
-pub const RING_BUFFER_SIZE: usize = (VOICE_SAMPLE_RATE * BUFFER_SECS) / EXPECTED_CHUNK_SIZE as usize;
+pub const RING_BUFFER_SIZE: usize = (VOICE_SAMPLE_RATE * BUFFER_SECS) / EXPECTED_CHUNK_SIZE;
 
 pub const MAX_DETECTION_DURATION_SECS: u128 = 80; // ~80ms is threshold for next detection run; based on chunk size
